+++ examples/app/src/test/kotlin/co/kofixtures/examples/Fixtures.kt
@@ -1,6 +1,0 @@
-package co.kofixtures.examples
-
-import co.kofixtures.ksp.KoFixture
-
-@KoFixture(packages = ["co.kofixtures.examples"])
-object Fixtures