+++ projects/kofixture-core/src/commonMain/kotlin/co/kofixtures/core/RegistryKey.kt
@@ -1,8 +1,0 @@
-package co.kofixtures.core
-
-import kotlin.reflect.KType
-
-data class RegistryKey(
-    val type: KType,
-    val tag: String? = null,
-)