+++ projects/kofixture-core/src/commonMain/kotlin/io/kofixture/core/ConcurrentMap.kt
@@ -1,0 +1,3 @@
+package io.kofixture.core
+
+internal expect fun <K : Any, V : Any> concurrentMutableMapOf(): MutableMap<K, V>