+++ projects/kofixture-core/src/commonMain/kotlin/io/kofixture/core/RegistryKey.kt
@@ -1,0 +1,8 @@
+package io.kofixture.core
+
+import kotlin.reflect.KType
+
+data class RegistryKey(
+    val type: KType,
+    val tag: String? = null,
+)