+++ projects/kofixture-core/src/jsMain/kotlin/co/kofixtures/core/ConcurrentMap.kt
@@ -1,3 +1,0 @@
-package co.kofixtures.core
-
-internal actual fun <K : Any, V : Any> concurrentMutableMapOf(): MutableMap<K, V> = mutableMapOf()