+++ projects/kofixture-core/src/jsMain/kotlin/io/kofixture/core/ConcurrentMap.kt
@@ -1,0 +1,3 @@
+package io.kofixture.core
+
+internal actual fun <K : Any, V : Any> concurrentMutableMapOf(): MutableMap<K, V> = mutableMapOf()