+++ projects/kofixture-core/src/jvmMain/kotlin/co/kofixtures/core/ConcurrentMap.kt
@@ -1,5 +1,0 @@
-package co.kofixtures.core
-
-import java.util.concurrent.ConcurrentHashMap
-
-internal actual fun <K : Any, V : Any> concurrentMutableMapOf(): MutableMap<K, V> = ConcurrentHashMap()